fix(domain): exact decimal arithmetic for quantities, prices and money
Quantities, prices, fees, money, holdings, FX rates, goal targets, the
price cache and daily-wealth totals were binary double end to end (and
REAL in SQLite), so balances showed artifacts like 0.13745907000000002
and 18-decimal token amounts could not be represented. They are now
System.Decimal (28-29 significant digits) from parsing through storage
(EF stores decimal as TEXT) to the API and the Blazor client models,
which deserialize the JSON number text exactly. Rounding happens only
at display; charts convert to double at the plotting boundary. Ratios
(percent changes) and volumes stay double.

Existing databases are migrated by the schema upgrader's table rebuild;
verified on a copy of a real pre-decimal database (every transaction's
quantity/price/fee preserved) via the opt-in CAPITRACK_LEGACY_DB test.
Format.Shares(decimal) now shows every significant digit, so dust such
as 0.000000001 no longer renders as 0.0000.

+++ src/client/Client.Application/Models/AccountSummaryDto.cs
@@ -5,7 +5,7 @@ public class AccountSummaryDto
 {
     public int AccountId { get; set; }
     public string AccountName { get; set; } = "";
-    public double MarketValue { get; set; }
-    public double CostBasis { get; set; }
+    public decimal MarketValue { get; set; }
+    public decimal CostBasis { get; set; }
     public int HoldingsCount { get; set; }
 }
