+++ src/client/Client.Application/Models/HistoryPointDto.cs
@@ -4,9 +4,9 @@ namespace Client.Application.Models;
 public class HistoryPointDto
 {
     public DateTime Date { get; set; }
-    public double? Close { get; set; }
-    public double? Open { get; set; }
-    public double? High { get; set; }
-    public double? Low { get; set; }
+    public decimal? Close { get; set; }
+    public decimal? Open { get; set; }
+    public decimal? High { get; set; }
+    public decimal? Low { get; set; }
     public double? Volume { get; set; }
 }
