+++ src/server/Server.Application/Prices/DailyWealthDto.cs
@@ -6,4 +6,4 @@ namespace Server.Application.Prices;
 /// <param name="TotalCost">The total cost basis on that date.</param>
 /// <param name="BaseCurrency">The base currency the totals are expressed in.</param>
 /// <param name="Details">An optional per-snapshot detail payload.</param>
-public record DailyWealthDto(string Date, double TotalWealth, double TotalCost, string BaseCurrency, object? Details);
+public record DailyWealthDto(string Date, decimal TotalWealth, decimal TotalCost, string BaseCurrency, object? Details);
