+++ src/server/Server.Application/Prices/DailyWealthSnapshotDto.cs
@@ -5,4 +5,4 @@ namespace Server.Application.Prices;
 /// <param name="TotalWealth">The total wealth on that date.</param>
 /// <param name="TotalCost">The total cost basis on that date.</param>
 /// <param name="BaseCurrency">The base currency the totals are expressed in.</param>
-public record DailyWealthSnapshotDto(string Date, double TotalWealth, double TotalCost, string BaseCurrency);
+public record DailyWealthSnapshotDto(string Date, decimal TotalWealth, decimal TotalCost, string BaseCurrency);
